apply the modified model
