--- src/main/java/vn/edu/hcmuaf/fit/webbanquanao/webpage/cart/controller/Show.java
@@ -1,21 +1,0 @@
-package vn.edu.hcmuaf.fit.webbanquanao.webpage.cart.controller;
-
-import jakarta.servlet.ServletException;
-import jakarta.servlet.annotation.WebServlet;
-import jakarta.servlet.http.HttpServlet;
-import jakarta.servlet.http.HttpServletRequest;
-import jakarta.servlet.http.HttpServletResponse;
-
-import java.io.IOException;
-
-@WebServlet(name="ShowCart", value="/ShowCart")
-public class Show extends HttpServlet {
-    @Override
-    protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        req.getRequestDispatcher("cart.jsp").forward(req, resp);
-    }
-
-    protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-
-    }
-}
