+++ src/main/java/vn/edu/hcmuaf/fit/webbanquanao/webpage/cart/model/CartItem.java
@@ -1,4 +1,4 @@
-package vn.edu.hcmuaf.fit.webbanquanao.webpage.newModel;
+package vn.edu.hcmuaf.fit.webbanquanao.webpage.cart.model;
 
 public class CartItem {
     private int id;
