+++ src/main/java/vn/edu/hcmuaf/fit/webbanquanao/webpage/product/dao/ProductDAO.java
@@ -116,283 +116,284 @@ public class ProductDAO {
 
     // Get products by query
     public Map<Integer, Product> getProductsByQuery(String query) {
-        return conn.get().withHandle(h -> {
-            Map<Integer, Product> result = new HashMap<>();
-            try (PreparedStatement stmt = h.getConnection().prepareStatement(query);
-                 ResultSet rs = stmt.executeQuery()
-            ) {
-                while (rs.next()) {
-                    int productId = rs.getInt("id");
-                    Product p = result.getOrDefault(productId, new Product());
-
-                    p.setId(productId);
-                    p.setType(rs.getString("type"));
-                    p.setCategory(rs.getString("category"));
-                    p.setSupplier(rs.getString("supplier"));
-                    p.setName(rs.getString("name"));
-                    p.setDescription(rs.getString("description"));
-                    p.setReleaseDate(rs.getDate("releaseDate"));
-                    p.setUnitSold(rs.getInt("unitSold"));
-                    p.setUnitPrice(rs.getDouble("unitPrice"));
-                    p.setStatus(rs.getBoolean("status"));
-                    p.setStock(rs.getInt("stock"));
-                    p.setProductDetailId(rs.getInt("productDetailId"));
-
-                    String size = rs.getString("size");
-                    String image = rs.getString("image");
-                    String color = rs.getString("color");
-
-                    if (!p.getSizes().contains(size))
-                        p.getSizes().add(size);
-                    if (!p.getImages().contains(image))
-                        p.getImages().add(image);
-                    if (!p.getColors().contains(color))
-                        p.getColors().add(color);
-
-                    result.put(productId, p);
-                }
-            } catch (SQLException e) {
-                e.printStackTrace();
-            }
-            return result;
-        });
+//        return conn.get().withHandle(h -> {
+//            Map<Integer, Product> result = new HashMap<>();
+//            try (PreparedStatement stmt = h.getConnection().prepareStatement(query);
+//                 ResultSet rs = stmt.executeQuery()
+//            ) {
+//                while (rs.next()) {
+//                    int productId = rs.getInt("id");
+//                    Product p = result.getOrDefault(productId, new Product());
+//
+//                    p.setId(productId);
+//                    p.setType(rs.getString("type"));
+//                    p.setCategory(rs.getString("category"));
+//                    p.setSupplier(rs.getString("supplier"));
+//                    p.setName(rs.getString("name"));
+//                    p.setDescription(rs.getString("description"));
+//                    p.setReleaseDate(rs.getDate("releaseDate"));
+//                    p.setUnitSold(rs.getInt("unitSold"));
+//                    p.setUnitPrice(rs.getDouble("unitPrice"));
+//                    p.setStatus(rs.getBoolean("status"));
+//                    p.setStock(rs.getInt("stock"));
+//                    p.setProductDetailId(rs.getInt("productDetailId"));
+//
+//                    String size = rs.getString("size");
+//                    String image = rs.getString("image");
+//                    String color = rs.getString("color");
+//
+//                    if (!p.getSizes().contains(size))
+//                        p.getSizes().add(size);
+//                    if (!p.getImages().contains(image))
+//                        p.getImages().add(image);
+//                    if (!p.getColors().contains(color))
+//                        p.getColors().add(color);
+//
+//                    result.put(productId, p);
+//                }
+//            } catch (SQLException e) {
+//                e.printStackTrace();
+//            }
+//            return result;
+//        });
+        Map<Integer, Product> result = new HashMap<>();
+        return result;
     }
 
     // Get list products by category
     public Map<Integer, Product> getProductsByCategory(String category) {
-        query = "SELECT" +
-                "   p.id AS id," +
-                "   t.name AS type," +
-                "   c.name AS category," +
-                "   s.supplierName AS supplier," +
-                "   p.productname AS name," +
-                "   p.description AS description," +
-                "   p.releaseDate AS releaseDate," +
-                "   p.unitSold AS unitSold," +
-                "   p.unitPrice AS unitPrice," +
-                "   p.status AS status," +
-                "   pd.size AS size," +
-                "   pd.stock AS stock," +
-                "   pd.image AS image," +
-                "   pd.color AS color, " +
-                "   pd.id AS productDetailId " +
-                "FROM products p " +
-                "JOIN product_details pd ON p.id = pd.productId " +
-                "JOIN categories c ON p.categoryId = c.id " +
-                "JOIN types t ON p.typeId = t.id " +
-                "JOIN suppliers s ON p.supplierId = s.id " +
-                "WHERE c.name = ?";
-
-        return conn.get().withHandle(h -> {
-            Map<Integer, Product> result = new HashMap<>();
-            try (PreparedStatement stmt = h.getConnection().prepareStatement(query)) {
-                stmt.setString(1, category);
-                try (ResultSet rs = stmt.executeQuery()) {
-                    while (rs.next()) {
-                        int productId = rs.getInt("id");
-                        Product p = result.getOrDefault(productId, new Product());
-
-                        p.setId(productId);
-                        p.setType(rs.getString("type"));
-                        p.setCategory(rs.getString("category"));
-                        p.setSupplier(rs.getString("supplier"));
-                        p.setName(rs.getString("name"));
-                        p.setDescription(rs.getString("description"));
-                        p.setReleaseDate(rs.getDate("releaseDate"));
-                        p.setUnitSold(rs.getInt("unitSold"));
-                        p.setUnitPrice(rs.getDouble("unitPrice"));
-                        p.setStatus(rs.getBoolean("status"));
-                        p.setStock(rs.getInt("stock"));
-
-                        String size = rs.getString("size");
-                        String image = rs.getString("image");
-                        String color = rs.getString("color");
-
-                        if (!p.getSizes().contains(size))
-                            p.getSizes().add(size);
-                        if (!p.getImages().contains(image))
-                            p.getImages().add(image);
-                        if (!p.getColors().contains(color))
-                            p.getColors().add(color);
-
-                        result.put(productId, p);
-                    }
-                }
-            } catch (SQLException e) {
-                e.printStackTrace();
-            }
-            return result;
-        });
+//        query = "SELECT" +
+//                "   p.id AS id," +
+//                "   t.name AS type," +
+//                "   c.name AS category," +
+//                "   s.supplierName AS supplier," +
+//                "   p.productname AS name," +
+//                "   p.description AS description," +
+//                "   p.releaseDate AS releaseDate," +
+//                "   p.unitSold AS unitSold," +
+//                "   p.unitPrice AS unitPrice," +
+//                "   p.status AS status," +
+//                "   pd.size AS size," +
+//                "   pd.stock AS stock," +
+//                "   pd.image AS image," +
+//                "   pd.color AS color, " +
+//                "   pd.id AS productDetailId " +
+//                "FROM products p " +
+//                "JOIN product_details pd ON p.id = pd.productId " +
+//                "JOIN categories c ON p.categoryId = c.id " +
+//                "JOIN types t ON p.typeId = t.id " +
+//                "JOIN suppliers s ON p.supplierId = s.id " +
+//                "WHERE c.name = ?";
+//
+//        return conn.get().withHandle(h -> {
+//            Map<Integer, Product> result = new HashMap<>();
+//            try (PreparedStatement stmt = h.getConnection().prepareStatement(query)) {
+//                stmt.setString(1, category);
+//                try (ResultSet rs = stmt.executeQuery()) {
+//                    while (rs.next()) {
+//                        int productId = rs.getInt("id");
+//                        Product p = result.getOrDefault(productId, new Product());
+//
+//                        p.setId(productId);
+//                        p.setType(rs.getString("type"));
+//                        p.setCategory(rs.getString("category"));
+//                        p.setSupplier(rs.getString("supplier"));
+//                        p.setName(rs.getString("name"));
+//                        p.setDescription(rs.getString("description"));
+//                        p.setReleaseDate(rs.getDate("releaseDate"));
+//                        p.setUnitSold(rs.getInt("unitSold"));
+//                        p.setUnitPrice(rs.getDouble("unitPrice"));
+//                        p.setStatus(rs.getBoolean("status"));
+//                        p.setStock(rs.getInt("stock"));
+//
+//                        String size = rs.getString("size");
+//                        String image = rs.getString("image");
+//                        String color = rs.getString("color");
+//
+//                        if (!p.getSizes().contains(size))
+//                            p.getSizes().add(size);
+//                        if (!p.getImages().contains(image))
+//                            p.getImages().add(image);
+//                        if (!p.getColors().contains(color))
+//                            p.getColors().add(color);
+//
+//                        result.put(productId, p);
+//                    }
+//                }
+//            } catch (SQLException e) {
+//                e.printStackTrace();
+//            }
+//            return result;
+//        });
+        Map<Integer, Product> result = new HashMap<>();
+        return result;
     }
 
     // Get filtered products
     public Map<Integer, Product> getFilteredProducts(String category, String[] type, String[] sizeF, Double priceMin, Double priceMax) {
-        query = "SELECT" +
-                "   p.id AS id," +
-                "   t.name AS type," +
-                "   c.name AS category," +
-                "   s.supplierName AS supplier," +
-                "   p.productname AS name," +
-                "   p.description AS description," +
-                "   p.releaseDate AS releaseDate," +
-                "   p.unitSold AS unitSold," +
-                "   p.unitPrice AS unitPrice," +
-                "   p.status AS status," +
-                "   pd.size AS size," +
-                "   pd.stock AS stock," +
-                "   pd.image AS image," +
-                "   pd.color AS color, " +
-                "   pd.id AS productDetailId " +
-                "FROM products p " +
-                "JOIN product_details pd ON p.id = pd.productId " +
-                "JOIN categories c ON p.categoryId = c.id " +
-                "JOIN types t ON p.typeId = t.id " +
-                "JOIN suppliers s ON p.supplierId = s.id " +
-                "WHERE c.name = ?";
-
-        if (type != null && type.length > 0) {
-            query += " AND t.name IN (" + String.join(",", Collections.nCopies(type.length, "?")) + ")";
-        }
-        if (sizeF != null && sizeF.length > 0) {
-            query += " AND pd.size IN (" + String.join(",", Collections.nCopies(sizeF.length, "?")) + ")";
-        }
-        if (priceMin != null) {
-            query += " AND p.unitPrice >= ?";
-        }
-        if (priceMax != null) {
-            query += " AND p.unitPrice <= ?";
-        }
-
-        return conn.get().withHandle(h -> {
-            Map<Integer, Product> result = new HashMap<>();
-            try (PreparedStatement stmt = h.getConnection().prepareStatement(query)) {
-                int paramIndex = 1;
-                stmt.setString(paramIndex++, category);
-                if (type != null && type.length > 0) {
-                    for (String t : type) {
-                        stmt.setString(paramIndex++, t);
-                    }
-                }
-                if (sizeF != null && sizeF.length > 0) {
-                    for (String size : sizeF) {
-                        stmt.setString(paramIndex++, size);
-                    }
-                }
-                if (priceMin != null) {
-                    stmt.setDouble(paramIndex++, priceMin);
-                }
-                if (priceMax != null) {
-                    stmt.setDouble(paramIndex, priceMax);
-                }
-
-                try (ResultSet rs = stmt.executeQuery()) {
-                    while (rs.next()) {
-                        int productId = rs.getInt("id");
-                        Product p = result.getOrDefault(productId, new Product());
-
-                        p.setId(productId);
-                        p.setType(rs.getString("type"));
-                        p.setCategory(rs.getString("category"));
-                        p.setSupplier(rs.getString("supplier"));
-                        p.setName(rs.getString("name"));
-                        p.setDescription(rs.getString("description"));
-                        p.setReleaseDate(rs.getDate("releaseDate"));
-                        p.setUnitSold(rs.getInt("unitSold"));
-                        p.setUnitPrice(rs.getDouble("unitPrice"));
-                        p.setStatus(rs.getBoolean("status"));
-                        p.setStock(rs.getInt("stock"));
-
-                        String size = rs.getString("size");
-                        String image = rs.getString("image");
-                        String color = rs.getString("color");
-
-                        if (!p.getSizes().contains(size))
-                            p.getSizes().add(size);
-                        if (!p.getImages().contains(image))
-                            p.getImages().add(image);
-                        if (!p.getColors().contains(color))
-                            p.getColors().add(color);
-
-                        result.put(productId, p);
-                    }
-                }
-            } catch (SQLException e) {
-                e.printStackTrace();
-            }
-            return result;
-        });
+//        query = "SELECT" +
+//                "   p.id AS id," +
+//                "   t.name AS type," +
+//                "   c.name AS category," +
+//                "   s.supplierName AS supplier," +
+//                "   p.productname AS name," +
+//                "   p.description AS description," +
+//                "   p.releaseDate AS releaseDate," +
+//                "   p.unitSold AS unitSold," +
+//                "   p.unitPrice AS unitPrice," +
+//                "   p.status AS status," +
+//                "   pd.size AS size," +
+//                "   pd.stock AS stock," +
+//                "   pd.image AS image," +
+//                "   pd.color AS color, " +
+//                "   pd.id AS productDetailId " +
+//                "FROM products p " +
+//                "JOIN product_details pd ON p.id = pd.productId " +
+//                "JOIN categories c ON p.categoryId = c.id " +
+//                "JOIN types t ON p.typeId = t.id " +
+//                "JOIN suppliers s ON p.supplierId = s.id " +
+//                "WHERE c.name = ?";
+//
+//        if (type != null && type.length > 0) {
+//            query += " AND t.name IN (" + String.join(",", Collections.nCopies(type.length, "?")) + ")";
+//        }
+//        if (sizeF != null && sizeF.length > 0) {
+//            query += " AND pd.size IN (" + String.join(",", Collections.nCopies(sizeF.length, "?")) + ")";
+//        }
+//        if (priceMin != null) {
+//            query += " AND p.unitPrice >= ?";
+//        }
+//        if (priceMax != null) {
+//            query += " AND p.unitPrice <= ?";
+//        }
+//
+//        return conn.get().withHandle(h -> {
+//            Map<Integer, Product> result = new HashMap<>();
+//            try (PreparedStatement stmt = h.getConnection().prepareStatement(query)) {
+//                int paramIndex = 1;
+//                stmt.setString(paramIndex++, category);
+//                if (type != null && type.length > 0) {
+//                    for (String t : type) {
+//                        stmt.setString(paramIndex++, t);
+//                    }
+//                }
+//                if (sizeF != null && sizeF.length > 0) {
+//                    for (String size : sizeF) {
+//                        stmt.setString(paramIndex++, size);
+//                    }
+//                }
+//                if (priceMin != null) {
+//                    stmt.setDouble(paramIndex++, priceMin);
+//                }
+//                if (priceMax != null) {
+//                    stmt.setDouble(paramIndex, priceMax);
+//                }
+//
+//                try (ResultSet rs = stmt.executeQuery()) {
+//                    while (rs.next()) {
+//                        int productId = rs.getInt("id");
+//                        Product p = result.getOrDefault(productId, new Product());
+//
+//                        p.setId(productId);
+//                        p.setType(rs.getString("type"));
+//                        p.setCategory(rs.getString("category"));
+//                        p.setSupplier(rs.getString("supplier"));
+//                        p.setName(rs.getString("name"));
+//                        p.setDescription(rs.getString("description"));
+//                        p.setReleaseDate(rs.getDate("releaseDate"));
+//                        p.setUnitSold(rs.getInt("unitSold"));
+//                        p.setUnitPrice(rs.getDouble("unitPrice"));
+//                        p.setStatus(rs.getBoolean("status"));
+//                        p.setStock(rs.getInt("stock"));
+//
+//                        String size = rs.getString("size");
+//                        String image = rs.getString("image");
+//                        String color = rs.getString("color");
+//
+//                        if (!p.getSizes().contains(size))
+//                            p.getSizes().add(size);
+//                        if (!p.getImages().contains(image))
+//                            p.getImages().add(image);
+//                        if (!p.getColors().contains(color))
+//                            p.getColors().add(color);
+//
+//                        result.put(productId, p);
+//                    }
+//                }
+//            } catch (SQLException e) {
+//                e.printStackTrace();
+//            }
+//            return result;
+//        });
+        Map<Integer, Product> result = new HashMap<>();
+        return result;
     }
 
     public Map<Integer, Product> searchByName(String name) {
-        String query = "SELECT" +
-                "   p.id AS id," +
-                "   t.name AS type," +
-                "   c.name AS category," +
-                "   s.supplierName AS supplier," +
-                "   p.productname AS name," +
-                "   p.description AS description," +
-                "   p.releaseDate AS releaseDate," +
-                "   p.unitSold AS unitSold," +
-                "   p.unitPrice AS unitPrice," +
-                "   p.status AS status," +
-                "   pd.size AS size," +
-                "   pd.stock AS stock," +
-                "   pd.image AS image," +
-                "   pd.color AS color, " +
-                "   pd.id AS productDetailId " +
-                "FROM products p " +
-                "JOIN product_details pd ON p.id = pd.productId " +
-                "JOIN categories c ON p.categoryId = c.id " +
-                "JOIN types t ON p.typeId = t.id " +
-                "JOIN suppliers s ON p.supplierId = s.id " +
-                "WHERE p.productname LIKE ?";
-
-        return conn.get().withHandle(h -> {
-            Map<Integer, Product> result = new HashMap<>();
-            try (PreparedStatement stmt = h.getConnection().prepareStatement(query)) {
-                stmt.setString(1, "%" + name + "%");
-                try (ResultSet rs = stmt.executeQuery()) {
-                    while (rs.next()) {
-                        int productId = rs.getInt("id");
-                        Product p = result.getOrDefault(productId, new Product());
-
-                        p.setId(productId);
-                        p.setType(rs.getString("type"));
-                        p.setCategory(rs.getString("category"));
-                        p.setSupplier(rs.getString("supplier"));
-                        p.setName(rs.getString("name"));
-                        p.setDescription(rs.getString("description"));
-                        p.setReleaseDate(rs.getDate("releaseDate"));
-                        p.setUnitSold(rs.getInt("unitSold"));
-                        p.setUnitPrice(rs.getDouble("unitPrice"));
-                        p.setStatus(rs.getBoolean("status"));
-                        p.setStock(rs.getInt("stock"));
-
-                        String size = rs.getString("size");
-                        String image = rs.getString("image");
-                        String color = rs.getString("color");
-
-                        if (!p.getSizes().contains(size))
-                            p.getSizes().add(size);
-                        if (!p.getImages().contains(image))
-                            p.getImages().add(image);
-                        if (!p.getColors().contains(color))
-                            p.getColors().add(color);
-
-                        result.put(productId, p);
-                    }
-                }
-            } catch (SQLException e) {
-                e.printStackTrace();
-            }
-            return result;
-        });
-    }
-
-    public static void main(String[] args) {
-        ProductDAO dao = new ProductDAO();
-//        System.out.println(dao.getFilteredProducts("Nam", new String[]{"Quần"}, new String[]{"M","L"}, null, 400000.0));
-//        System.out.println(dao.searchByName("Quần tây"));
-        System.out.println(dao.getAllProducts());
+//        String query = "SELECT" +
+//                "   p.id AS id," +
+//                "   t.name AS type," +
+//                "   c.name AS category," +
+//                "   s.supplierName AS supplier," +
+//                "   p.productname AS name," +
+//                "   p.description AS description," +
+//                "   p.releaseDate AS releaseDate," +
+//                "   p.unitSold AS unitSold," +
+//                "   p.unitPrice AS unitPrice," +
+//                "   p.status AS status," +
+//                "   pd.size AS size," +
+//                "   pd.stock AS stock," +
+//                "   pd.image AS image," +
+//                "   pd.color AS color, " +
+//                "   pd.id AS productDetailId " +
+//                "FROM products p " +
+//                "JOIN product_details pd ON p.id = pd.productId " +
+//                "JOIN categories c ON p.categoryId = c.id " +
+//                "JOIN types t ON p.typeId = t.id " +
+//                "JOIN suppliers s ON p.supplierId = s.id " +
+//                "WHERE p.productname LIKE ?";
+//
+//        return conn.get().withHandle(h -> {
+//            Map<Integer, Product> result = new HashMap<>();
+//            try (PreparedStatement stmt = h.getConnection().prepareStatement(query)) {
+//                stmt.setString(1, "%" + name + "%");
+//                try (ResultSet rs = stmt.executeQuery()) {
+//                    while (rs.next()) {
+//                        int productId = rs.getInt("id");
+//                        Product p = result.getOrDefault(productId, new Product());
+//
+//                        p.setId(productId);
+//                        p.setType(rs.getString("type"));
+//                        p.setCategory(rs.getString("category"));
+//                        p.setSupplier(rs.getString("supplier"));
+//                        p.setName(rs.getString("name"));
+//                        p.setDescription(rs.getString("description"));
+//                        p.setReleaseDate(rs.getDate("releaseDate"));
+//                        p.setUnitSold(rs.getInt("unitSold"));
+//                        p.setUnitPrice(rs.getDouble("unitPrice"));
+//                        p.setStatus(rs.getBoolean("status"));
+//                        p.setStock(rs.getInt("stock"));
+//
+//                        String size = rs.getString("size");
+//                        String image = rs.getString("image");
+//                        String color = rs.getString("color");
+//
+//                        if (!p.getSizes().contains(size))
+//                            p.getSizes().add(size);
+//                        if (!p.getImages().contains(image))
+//                            p.getImages().add(image);
+//                        if (!p.getColors().contains(color))
+//                            p.getColors().add(color);
+//
+//                        result.put(productId, p);
+//                    }
+//                }
+//            } catch (SQLException e) {
+//                e.printStackTrace();
+//            }
+//            return result;
+//        });
+        Map<Integer, Product> result = new HashMap<>();
+        return result;
     }
 }
